+++ Realm/ObjectServerTests/SwiftUIServerTests.swift
@@ -30,6 +30,12 @@ import RealmSyncTestSupport
 @MainActor
 class SwiftUIServerTests: SwiftSyncTestCase {
 
+    enum OpenType {
+        case configuration(Realm.Configuration)
+        case pbs(String)
+        case flexibleSync
+    }
+
     // Configuration for tests
     private func configuration<T: BSON>(user: User, partition: T) -> Realm.Configuration {
         var userConfiguration = user.configuration(partitionValue: partition)
@@ -46,13 +52,23 @@ class SwiftUIServerTests: SwiftSyncTestCase {
     var cancellables: Set<AnyCancellable> = []
 
     // MARK: - AsyncOpen
-    func asyncOpen<T: BSON>(user: User, appId: String? = nil, partitionValue: T, timeout: UInt? = nil,
-                            handler: @escaping (AsyncOpenState) -> Void) {
-        let configuration = self.configuration(user: user, partition: partitionValue)
-        let asyncOpen = AsyncOpen(appId: appId,
-                                  partitionValue: partitionValue,
+    func asyncOpen(appId: String? = nil, openType: OpenType, timeout: UInt? = nil,
+                   handler: @escaping (AsyncOpenState) -> Void) {
+        let asyncOpen: AsyncOpen
+        switch openType {
+        case .configuration(let configuration):
+            asyncOpen = AsyncOpen(appId: appId,
                                   configuration: configuration,
                                   timeout: timeout)
+        case .pbs(let partitionValue):
+            asyncOpen = AsyncOpen(appId: appId,
+                                  partitionValue: partitionValue,
+                                  timeout: timeout)
+        case .flexibleSync:
+            asyncOpen = AsyncOpen(appId: appId,
+                                  timeout: timeout)
+        }
+
         _ = asyncOpen.wrappedValue // Retrieving the wrappedValue to simulate a SwiftUI environment where this is called when initialising the view.
         asyncOpen.projectedValue
             .sink(receiveValue: handler)
@@ -62,10 +78,10 @@ class SwiftUIServerTests: SwiftSyncTestCase {
     }
 
     func testAsyncOpenOpenRealm() throws {
-        let user = try logInUser(for: basicCredentials())
+        _ = try logInUser(for: basicCredentials())
 
         let ex = expectation(description: "download-realm-async-open")
-        asyncOpen(user: user, appId: appId, partitionValue: #function) { asyncOpenState in
+        asyncOpen(appId: appId, openType: .pbs(#function)) { asyncOpenState in
             if case let .open(realm) = asyncOpenState {
                 XCTAssertNotNil(realm)
                 ex.fulfill()
@@ -81,7 +97,7 @@ class SwiftUIServerTests: SwiftSyncTestCase {
         executeChild()
 
         let ex = expectation(description: "download-populated-realm-async-open")
-        asyncOpen(user: user, appId: appId, partitionValue: #function) { asyncOpenState in
+        asyncOpen(appId: appId, openType: .pbs(#function)) { asyncOpenState in
             if case let .open(realm) = asyncOpenState {
                 XCTAssertNotNil(realm)
                 self.checkCount(expected: SwiftSyncTestCase.bigObjectCount, realm, SwiftHugeSyncObject.self)
@@ -95,7 +111,7 @@ class SwiftUIServerTests: SwiftSyncTestCase {
         user.logOut { _ in } // Logout current user
 
         let ex = expectation(description: "download-realm-async-open-not-logged")
-        asyncOpen(user: user, appId: appId, partitionValue: #function) { asyncOpenState in
+        asyncOpen(appId: appId, openType: .pbs(#function)) { asyncOpenState in
             if case .waitingForUser = asyncOpenState {
                 ex.fulfill()
             }
@@ -111,11 +127,11 @@ class SwiftUIServerTests: SwiftSyncTestCase {
         let appConfig = AppConfiguration(baseURL: "http://localhost:5678",
                                          transport: AsyncOpenConnectionTimeoutTransport())
         let app = App(id: appId, configuration: appConfig)
-        let user = try logInUser(for: basicCredentials(app: app), app: app)
+        _ = try logInUser(for: basicCredentials(app: app), app: app)
 
         proxy.dropConnections = true
         let ex = expectation(description: "download-realm-async-open-no-connection")
-        asyncOpen(user: user, appId: appId, partitionValue: #function, timeout: 1000) { asyncOpenState in
+        asyncOpen(appId: appId, openType: .pbs(#function), timeout: 1000) { asyncOpenState in
             if case let .error(error) = asyncOpenState,
                let nsError = error as NSError? {
                 XCTAssertEqual(nsError.code, Int(ETIMEDOUT))
@@ -137,7 +153,7 @@ class SwiftUIServerTests: SwiftSyncTestCase {
         executeChild()
 
         let ex = expectation(description: "progress-async-open")
-        asyncOpen(user: user, appId: appId, partitionValue: #function) { asyncOpenState in
+        asyncOpen(appId: appId, openType: .pbs(#function)) { asyncOpenState in
             if case let .progress(progress) = asyncOpenState {
                 XCTAssertTrue(progress.fractionCompleted > 0)
                 if progress.isFinished {
@@ -156,7 +172,7 @@ class SwiftUIServerTests: SwiftSyncTestCase {
         executeChild()
 
         let ex = expectation(description: "download-cached-app-async-open")
-        asyncOpen(user: user, partitionValue: #function) { asyncOpenState in
+        asyncOpen(openType: .pbs(#function)) { asyncOpenState in
             if case let .open(realm) = asyncOpenState {
                 XCTAssertNotNil(realm)
                 self.checkCount(expected: SwiftSyncTestCase.bigObjectCount, realm, SwiftHugeSyncObject.self)
@@ -189,7 +205,7 @@ class SwiftUIServerTests: SwiftSyncTestCase {
         executeChild()
 
         let ex = expectation(description: "download-partition-value-async-open")
-        asyncOpen(user: user, partitionValue: partitionValueA) { asyncOpenState in
+        asyncOpen(openType: .pbs(partitionValueA)) { asyncOpenState in
             if case let .open(realm) = asyncOpenState {
                 XCTAssertNotNil(realm)
                 self.checkCount(expected: SwiftSyncTestCase.bigObjectCount, realm, SwiftHugeSyncObject.self)
@@ -198,7 +214,7 @@ class SwiftUIServerTests: SwiftSyncTestCase {
         }
 
         let ex2 = expectation(description: "download-other-partition-value-async-open")
-        asyncOpen(user: user, partitionValue: partitionValueB) { asyncOpenState in
+        asyncOpen(openType: .pbs(partitionValueB)) { asyncOpenState in
             if case let .open(realm) = asyncOpenState {
                 XCTAssertNotNil(realm)
                 self.checkCount(expected: 0, realm, SwiftHugeSyncObject.self)
@@ -222,7 +238,7 @@ class SwiftUIServerTests: SwiftSyncTestCase {
         executeChild()
 
         let ex = expectation(description: "test-multiuser1-app-async-open")
-        asyncOpen(user: syncUser2, appId: appId, partitionValue: partitionValueB) { asyncOpenState in
+        asyncOpen(appId: appId, openType: .pbs(partitionValueB)) { asyncOpenState in
             if case let .open(realm) = asyncOpenState {
                 XCTAssertNotNil(realm)
                 self.checkCount(expected: 0, realm, SwiftHugeSyncObject.self)
@@ -235,7 +251,7 @@ class SwiftUIServerTests: SwiftSyncTestCase {
         XCTAssertEqual(syncUser1.id, app.currentUser!.id)
 
         let ex2 = expectation(description: "test-multiuser2-app-async-open")
-        asyncOpen(user: syncUser2, appId: appId, partitionValue: partitionValueA) { asyncOpenState in
+        asyncOpen(appId: appId, openType: .pbs(partitionValueA)) { asyncOpenState in
             if case let .open(realm) = asyncOpenState {
                 XCTAssertNotNil(realm)
                 self.checkCount(expected: SwiftSyncTestCase.bigObjectCount, realm, SwiftHugeSyncObject.self)
@@ -254,9 +270,9 @@ class SwiftUIServerTests: SwiftSyncTestCase {
         }
         executeChild()
 
-        let anonymousUser = try logInUser(for: .anonymous)
+        _ = try logInUser(for: .anonymous)
         let ex = expectation(description: "download-realm-anonymous-user-async-open")
-        asyncOpen(user: anonymousUser, appId: appId, partitionValue: partitionValueA) { asyncOpenState in
+        asyncOpen(appId: appId, openType: .pbs(partitionValueA)) { asyncOpenState in
             if case let .open(realm) = asyncOpenState {
                 XCTAssertNotNil(realm)
                 self.checkCount(expected: 0, realm, SwiftHugeSyncObject.self)
@@ -267,7 +283,7 @@ class SwiftUIServerTests: SwiftSyncTestCase {
         app.currentUser?.logOut { _ in } // Logout anonymous user
 
         let ex2 = expectation(description: "download-realm-after-logout-async-open")
-        asyncOpen(user: user, appId: appId, partitionValue: partitionValueB) { asyncOpenState in
+        asyncOpen(appId: appId, openType: .pbs(partitionValueB)) { asyncOpenState in
             if case let .open(realm) = asyncOpenState {
                 XCTAssertNotNil(realm)
                 self.checkCount(expected: SwiftSyncTestCase.bigObjectCount, realm, SwiftHugeSyncObject.self)
@@ -276,13 +292,137 @@ class SwiftUIServerTests: SwiftSyncTestCase {
         }
     }
 
+    func testAsyncOpenFlexibleSyncInit() throws {
+        try populateFlexibleSyncData { realm in
+            for i in 1...10 {
+                // Using firstname to query only objects from this test
+                let person = SwiftPerson(firstName: "\(#function)",
+                                         lastName: "lastname_\(i)",
+                                         age: i)
+                realm.add(person)
+            }
+        }
+
+        _ = try logInUser(for: basicCredentials(app: flexibleSyncApp), app: flexibleSyncApp)
+
+        let ex = expectation(description: "download-realm-flexible-async-open")
+        asyncOpen(appId: flexibleSyncAppId, openType: .flexibleSync, timeout: 1000) { asyncOpenState in
+            if case let .open(realm) = asyncOpenState {
+                XCTAssertNotNil(realm)
+                XCTAssertTrue(realm.isEmpty) // should not have downloaded anything, because there are no subscriptions
+                ex.fulfill()
+            }
+        }
+    }
+
+    func testAsyncOpenFlexibleSyncConfiguration() throws {
+        try populateFlexibleSyncData { realm in
+            for i in 1...10 {
+                // Using firstname to query only objects from this test
+                let person = SwiftPerson(firstName: "\(#function)",
+                                         lastName: "lastname_\(i)",
+                                         age: i)
+                realm.add(person)
+            }
+        }
+
+        let user = try logInUser(for: basicCredentials(app: flexibleSyncApp), app: flexibleSyncApp)
+
+        var configuration = user.flexibleSyncConfiguration(initialSubscriptions: { subs in
+            subs.append(QuerySubscription<SwiftPerson> {
+                $0.firstName == "\(#function)" && $0.age > 0
+            })
+        })
+        configuration.objectTypes = [SwiftPerson.self]
+
+        let ex = expectation(description: "download-realm-flexible-async-open")
+        asyncOpen(appId: flexibleSyncAppId, openType: .configuration(configuration), timeout: 1000) { asyncOpenState in
+            if case let .open(realm) = asyncOpenState {
+                XCTAssertNotNil(realm)
+                self.checkCount(expected: 10, realm, SwiftPerson.self)
+                ex.fulfill()
+            }
+        }
+    }
+
+    func testAsyncOpenPBSWithConfiguration() throws {
+        let user = try logInUser(for: basicCredentials())
+        if !isParent {
+            try populateRealm(user: user, partitionValue: #function)
+            return
+        }
+        executeChild()
+
+        var configuration = user.configuration(partitionValue: #function)
+        configuration.objectTypes = [SwiftHugeSyncObject.self]
+
+        let ex = expectation(description: "download-populated-realm-async-open")
+        asyncOpen(appId: appId, openType: .configuration(configuration)) { asyncOpenState in
+            if case let .open(realm) = asyncOpenState {
+                XCTAssertNotNil(realm)
+                self.checkCount(expected: SwiftSyncTestCase.bigObjectCount, realm, SwiftHugeSyncObject.self)
+                ex.fulfill()
+            }
+        }
+    }
+
+    func testAsyncOpenPBSWithClientReset() throws {
+        let user = try logInUser(for: basicCredentials())
+        var configuration = user.configuration(partitionValue: #function, clientResetMode: .manual(errorHandler: { _, _ in }))
+        configuration.objectTypes = [SwiftPerson.self]
+
+        let ex = expectation(description: "download-populated-realm-async-open")
+        asyncOpen(appId: appId, openType: .configuration(configuration)) { asyncOpenState in
+            if case let .open(realm) = asyncOpenState {
+                XCTAssertNotNil(realm)
+                switch configuration.syncConfiguration!.clientResetMode {
+                case .manual(let block):
+                    XCTAssertNotNil(block)
+                default:
+                    XCTFail("Should be set to manual")
+                }
+                ex.fulfill()
+            }
+        }
+    }
+
+    func testAsyncOpenFlexibleSyncWithClientReset() throws {
+        let user = try logInUser(for: basicCredentials(app: flexibleSyncApp), app: flexibleSyncApp)
+        var configuration = user.flexibleSyncConfiguration(clientResetMode: .manual(errorHandler: { _, _ in }))
+        configuration.objectTypes = [SwiftPerson.self]
+
+        let ex = expectation(description: "download-populated-realm-async-open")
+        asyncOpen(appId: flexibleSyncAppId, openType: .configuration(configuration)) { asyncOpenState in
+            if case let .open(realm) = asyncOpenState {
+                XCTAssertNotNil(realm)
+                switch configuration.syncConfiguration!.clientResetMode {
+                case .manual(let block):
+                    XCTAssertNotNil(block)
+                default:
+                    XCTFail("Should be set to manual")
+                }
+                ex.fulfill()
+            }
+        }
+    }
+
     // MARK: - AutoOpen
-    func autoOpen(user: User, appId: String? = nil, partitionValue: String, timeout: UInt? = nil, handler: @escaping (AsyncOpenState) -> Void) {
-        let configuration = self.configuration(user: user, partition: partitionValue)
-        let autoOpen = AutoOpen(appId: appId,
-                                partitionValue: partitionValue,
+    func autoOpen(appId: String? = nil, openType: OpenType, timeout: UInt? = nil, handler: @escaping (AsyncOpenState) -> Void) {
+        let autoOpen: AutoOpen
+        switch openType {
+        case .configuration(let configuration):
+            autoOpen = AutoOpen(appId: appId,
                                 configuration: configuration,
                                 timeout: timeout)
+        case .pbs(let partitionValue):
+            autoOpen = AutoOpen(appId: appId,
+                                partitionValue: partitionValue,
+                                timeout: timeout)
+        case .flexibleSync:
+            autoOpen = AutoOpen(appId: appId,
+                                timeout: timeout)
+        }
+
         _ = autoOpen.wrappedValue // Retrieving the wrappedValue to simulate a SwiftUI environment where this is called when initialising the view.
         autoOpen.projectedValue
             .sink(receiveValue: handler)
@@ -292,10 +432,10 @@ class SwiftUIServerTests: SwiftSyncTestCase {
     }
 
     func testAutoOpenOpenRealm() throws {
-        let user = try logInUser(for: basicCredentials())
+        _ = try logInUser(for: basicCredentials())
 
         let ex = expectation(description: "download-realm-auto-open")
-        autoOpen(user: user, appId: appId, partitionValue: #function) { autoOpenState in
+        autoOpen(appId: appId, openType: .pbs(#function)) { autoOpenState in
             if case let .open(realm) = autoOpenState {
                 XCTAssertNotNil(realm)
                 ex.fulfill()
@@ -312,7 +452,7 @@ class SwiftUIServerTests: SwiftSyncTestCase {
         executeChild()
 
         let ex = expectation(description: "download-populated-realm-auto-open")
-        autoOpen(user: user, appId: appId, partitionValue: #function) { autoOpenState in
+        autoOpen(appId: appId, openType: .pbs(#function)) { autoOpenState in
             if case let .open(realm) = autoOpenState {
                 XCTAssertNotNil(realm)
                 self.checkCount(expected: SwiftSyncTestCase.bigObjectCount, realm, SwiftHugeSyncObject.self)
@@ -327,7 +467,7 @@ class SwiftUIServerTests: SwiftSyncTestCase {
         user.logOut { _ in } // Logout current user
 
         let ex = expectation(description: "download-realm-auto-open-not-logged")
-        autoOpen(user: user, appId: appId, partitionValue: #function) { autoOpenState in
+        autoOpen(appId: appId, openType: .pbs(#function)) { autoOpenState in
             if case .waitingForUser = autoOpenState {
                 ex.fulfill()
             }
@@ -347,10 +487,10 @@ class SwiftUIServerTests: SwiftSyncTestCase {
         let appConfig = AppConfiguration(baseURL: "http://localhost:5678",
                                          transport: AsyncOpenConnectionTimeoutTransport())
         let app = App(id: appId, configuration: appConfig)
-        let user = try logInUser(for: basicCredentials(app: app), app: app)
+        _ = try logInUser(for: basicCredentials(app: app), app: app)
         proxy.dropConnections = true
         let ex = expectation(description: "download-realm-auto-open-no-connection")
-        autoOpen(user: user, appId: appId, partitionValue: #function, timeout: 1000) { autoOpenState in
+        autoOpen(appId: appId, openType: .pbs(#function), timeout: 1000) { autoOpenState in
             if case let .open(realm) = autoOpenState {
                 XCTAssertTrue(realm.isEmpty) // should not have downloaded anything
                 ex.fulfill()
@@ -382,27 +522,16 @@ class SwiftUIServerTests: SwiftSyncTestCase {
         let appConfig = AppConfiguration(baseURL: "http://localhost:5678",
                                          transport: AsyncOpenConnectionTimeoutTransport())
         let app = App(id: flexibleSyncAppId, configuration: appConfig)
-
-        let user = try logInUser(for: basicCredentials(app: app), app: app)
-        var configuration = user.flexibleSyncConfiguration()
-        configuration.objectTypes = [SwiftPerson.self]
+        _ = try logInUser(for: basicCredentials(app: app), app: app)
 
         proxy.dropConnections = true
         let ex = expectation(description: "download-realm-flexible-auto-open-no-connection")
-        let autoOpen = AutoOpen(appId: flexibleSyncAppId, configuration: configuration, timeout: 1000)
-
-        _ = autoOpen.wrappedValue // Retrieving the wrappedValue to simulate a SwiftUI environment where this is called when initialising the view.
-        autoOpen.projectedValue
-            .sink { autoOpenState in
-                if case let .open(realm) = autoOpenState {
-                    XCTAssertTrue(realm.isEmpty) // should not have downloaded anything
-                    ex.fulfill()
-                }
+        autoOpen(appId: flexibleSyncAppId, openType: .flexibleSync, timeout: 1000) { autoOpenState in
+            if case let .open(realm) = autoOpenState {
+                XCTAssertTrue(realm.isEmpty) // should not have downloaded anything
+                ex.fulfill()
             }
-            .store(in: &cancellables)
-
-        waitForExpectations(timeout: 10.0)
-        autoOpen.cancel()
+        }
 
         App.resetAppCache()
         proxy.stop()
@@ -414,9 +543,9 @@ class SwiftUIServerTests: SwiftSyncTestCase {
             try populateRealm(user: user, partitionValue: #function)
         }
 
-        let user = try logInUser(for: basicCredentials())
+        _ = try logInUser(for: basicCredentials())
         let ex = expectation(description: "progress-auto-open")
-        autoOpen(user: user, appId: appId, partitionValue: #function) { autoOpenState in
+        autoOpen(appId: appId, openType: .pbs(#function)) { autoOpenState in
             if case let .progress(progress) = autoOpenState {
                 XCTAssertTrue(progress.fractionCompleted > 0)
                 if progress.isFinished {
@@ -435,7 +564,7 @@ class SwiftUIServerTests: SwiftSyncTestCase {
         executeChild()
 
         let ex = expectation(description: "download-cached-app-auto-open")
-        autoOpen(user: user, partitionValue: #function) { autoOpenState in
+        autoOpen(openType: .pbs(#function)) { autoOpenState in
             if case let .open(realm) = autoOpenState {
                 XCTAssertNotNil(realm)
                 self.checkCount(expected: SwiftSyncTestCase.bigObjectCount, realm, SwiftHugeSyncObject.self)
@@ -474,7 +603,7 @@ class SwiftUIServerTests: SwiftSyncTestCase {
         executeChild()
 
         let ex = expectation(description: "test-multiuser1-app-auto-open")
-        autoOpen(user: syncUser2, appId: appId, partitionValue: partitionValueB) { autoOpenState in
+        autoOpen(appId: appId, openType: .pbs(partitionValueB)) { autoOpenState in
             if case let .open(realm) = autoOpenState {
                 XCTAssertNotNil(realm)
                 self.checkCount(expected: 0, realm, SwiftHugeSyncObject.self)
@@ -487,7 +616,7 @@ class SwiftUIServerTests: SwiftSyncTestCase {
         XCTAssertEqual(syncUser1.id, app.currentUser!.id)
 
         let ex2 = expectation(description: "test-multiuser2-app-auto-open")
-        autoOpen(user: syncUser1, appId: appId, partitionValue: partitionValueA) { autoOpenState in
+        autoOpen(appId: appId, openType: .pbs(partitionValueA)) { autoOpenState in
             if case let .open(realm) = autoOpenState {
                 XCTAssertNotNil(realm)
                 self.checkCount(expected: SwiftSyncTestCase.bigObjectCount, realm, SwiftHugeSyncObject.self)
@@ -500,9 +629,9 @@ class SwiftUIServerTests: SwiftSyncTestCase {
         let partitionValueA = #function
         let partitionValueB = "\(#function)bar"
 
-        let anonymousUser = try logInUser(for: .anonymous)
+        _ = try logInUser(for: .anonymous)
         let ex = expectation(description: "download-realm-anonymous-user-auto-open")
-        autoOpen(user: anonymousUser, appId: appId, partitionValue: partitionValueA) { autoOpenState in
+        autoOpen(appId: appId, openType: .pbs(partitionValueA)) { autoOpenState in
             if case let .open(realm) = autoOpenState {
                 XCTAssertNotNil(realm)
                 self.checkCount(expected: 0, realm, SwiftHugeSyncObject.self)
@@ -519,7 +648,7 @@ class SwiftUIServerTests: SwiftSyncTestCase {
         executeChild()
 
         let ex2 = expectation(description: "download-realm-after-logout-auto-open")
-        autoOpen(user: user, appId: appId, partitionValue: partitionValueB) { autoOpenState in
+        autoOpen(appId: appId, openType: .pbs(partitionValueB)) { autoOpenState in
             if case let .open(realm) = autoOpenState {
                 XCTAssertNotNil(realm)
                 self.checkCount(expected: SwiftSyncTestCase.bigObjectCount, realm, SwiftHugeSyncObject.self)
@@ -539,7 +668,7 @@ class SwiftUIServerTests: SwiftSyncTestCase {
         executeChild()
 
         let ex = expectation(description: "download-partition-value-auto-open")
-        autoOpen(user: user, partitionValue: partitionValueA) { autoOpenState in
+        autoOpen(openType: .pbs(partitionValueA)) { autoOpenState in
             if case let .open(realm) = autoOpenState {
                 XCTAssertNotNil(realm)
                 self.checkCount(expected: SwiftSyncTestCase.bigObjectCount, realm, SwiftHugeSyncObject.self)
@@ -548,7 +677,7 @@ class SwiftUIServerTests: SwiftSyncTestCase {
         }
 
         let ex2 = expectation(description: "download-other-partition-value-auto-open")
-        autoOpen(user: user, partitionValue: partitionValueB) { autoOpenState in
+        autoOpen(openType: .pbs(partitionValueB)) { autoOpenState in
             if case let .open(realm) = autoOpenState {
                 XCTAssertNotNil(realm)
                 self.checkCount(expected: 0, realm, SwiftHugeSyncObject.self)
@@ -569,7 +698,7 @@ class SwiftUIServerTests: SwiftSyncTestCase {
         executeChild()
 
         let ex = expectation(description: "download-partition-value-async-open-mixed")
-        asyncOpen(user: user, partitionValue: partitionValueA) { asyncOpenState in
+        asyncOpen(openType: .pbs(partitionValueA)) { asyncOpenState in
             if case let .open(realm) = asyncOpenState {
                 XCTAssertNotNil(realm)
                 self.checkCount(expected: SwiftSyncTestCase.bigObjectCount, realm, SwiftHugeSyncObject.self)
@@ -578,7 +707,7 @@ class SwiftUIServerTests: SwiftSyncTestCase {
         }
 
         let ex2 = expectation(description: "download-partition-value-auto-open-mixed")
-        autoOpen(user: user, partitionValue: partitionValueB) { autoOpenState in
+        autoOpen(openType: .pbs(partitionValueB)) { autoOpenState in
             if case let .open(realm) = autoOpenState {
                 XCTAssertNotNil(realm)
                 self.checkCount(expected: 0, realm, SwiftHugeSyncObject.self)
@@ -602,7 +731,7 @@ class SwiftUIServerTests: SwiftSyncTestCase {
         executeChild()
 
         let ex = expectation(description: "test-combine-multiuser1-app-auto-open")
-        autoOpen(user: syncUser2, appId: appId, partitionValue: partitionValueB) { autoOpenState in
+        autoOpen(appId: appId, openType: .pbs(partitionValueB)) { autoOpenState in
             if case let .open(realm) = autoOpenState {
                 XCTAssertNotNil(realm)
                 self.checkCount(expected: 0, realm, SwiftHugeSyncObject.self)
@@ -615,11 +744,125 @@ class SwiftUIServerTests: SwiftSyncTestCase {
         XCTAssertEqual(syncUser1.id, app.currentUser!.id)
 
         let ex2 = expectation(description: "test-combine-multiuser2-app-auto-open")
-        asyncOpen(user: syncUser1, appId: appId, partitionValue: partitionValueA) { asyncOpenState in
+        asyncOpen(appId: appId, openType: .pbs(partitionValueA)) { asyncOpenState in
             if case let .open(realm) = asyncOpenState {
                 XCTAssertNotNil(realm)
                 self.checkCount(expected: SwiftSyncTestCase.bigObjectCount, realm, SwiftHugeSyncObject.self)
                 ex2.fulfill()
+            }
+        }
+    }
+
+    func testAutoOpenFlexibleSyncInit() throws {
+        try populateFlexibleSyncData { realm in
+            for i in 1...10 {
+                // Using firstname to query only objects from this test
+                let person = SwiftPerson(firstName: "\(#function)",
+                                         lastName: "lastname_\(i)",
+                                         age: i)
+                realm.add(person)
+            }
+        }
+
+        _ = try logInUser(for: basicCredentials(app: flexibleSyncApp), app: flexibleSyncApp)
+
+        let ex = expectation(description: "download-realm-flexible-async-open")
+        autoOpen(appId: flexibleSyncAppId, openType: .flexibleSync, timeout: 1000) { autoOpenState in
+            if case let .open(realm) = autoOpenState {
+                XCTAssertNotNil(realm)
+                XCTAssertTrue(realm.isEmpty) // should not have downloaded anything, because there are no subscriptions
+                ex.fulfill()
+            }
+        }
+    }
+
+    func testAutoOpenFlexibleSyncConfiguration() throws {
+        try populateFlexibleSyncData { realm in
+            for i in 1...10 {
+                // Using firstname to query only objects from this test
+                let person = SwiftPerson(firstName: "\(#function)",
+                                         lastName: "lastname_\(i)",
+                                         age: i)
+                realm.add(person)
+            }
+        }
+
+        let user = try logInUser(for: basicCredentials(app: flexibleSyncApp), app: flexibleSyncApp)
+
+        var configuration = user.flexibleSyncConfiguration(initialSubscriptions: { subs in
+            subs.append(QuerySubscription<SwiftPerson> {
+                $0.firstName == "\(#function)" && $0.age > 0
+            })
+        })
+        configuration.objectTypes = [SwiftPerson.self]
+
+        let ex = expectation(description: "download-realm-flexible-async-open")
+        autoOpen(appId: flexibleSyncAppId, openType: .configuration(configuration), timeout: 1000) { autoOpenState in
+            if case let .open(realm) = autoOpenState {
+                XCTAssertNotNil(realm)
+                self.checkCount(expected: 10, realm, SwiftPerson.self)
+                ex.fulfill()
+            }
+        }
+    }
+
+    func testAutoOpenPBSWithConfiguration() throws {
+        let user = try logInUser(for: basicCredentials())
+        if !isParent {
+            try populateRealm(user: user, partitionValue: #function)
+            return
+        }
+        executeChild()
+
+        var configuration = user.configuration(partitionValue: #function)
+        configuration.objectTypes = [SwiftHugeSyncObject.self]
+
+        let ex = expectation(description: "download-populated-realm-async-open")
+        autoOpen(appId: appId, openType: .configuration(configuration)) { autoOpenState in
+            if case let .open(realm) = autoOpenState {
+                XCTAssertNotNil(realm)
+                self.checkCount(expected: SwiftSyncTestCase.bigObjectCount, realm, SwiftHugeSyncObject.self)
+                ex.fulfill()
+            }
+        }
+    }
+
+    func testAutoOpenPBSWithClientReset() throws {
+        let user = try logInUser(for: basicCredentials())
+        var configuration = user.configuration(partitionValue: #function, clientResetMode: .manual(errorHandler: { _, _ in }))
+        configuration.objectTypes = [SwiftPerson.self]
+
+        let ex = expectation(description: "download-populated-realm-async-open")
+        autoOpen(appId: appId, openType: .configuration(configuration)) { autoOpenState in
+            if case let .open(realm) = autoOpenState {
+                XCTAssertNotNil(realm)
+                switch configuration.syncConfiguration!.clientResetMode {
+                case .manual(let block):
+                    XCTAssertNotNil(block)
+                default:
+                    XCTFail("Should be set to manual")
+                }
+                ex.fulfill()
+            }
+        }
+    }
+
+    func testAutoOpenFlexibleSyncWithClientReset() throws {
+        let user = try logInUser(for: basicCredentials(app: flexibleSyncApp), app: flexibleSyncApp)
+        var configuration = user.flexibleSyncConfiguration(clientResetMode: .manual(errorHandler: { _, _ in }))
+        configuration.objectTypes = [SwiftPerson.self]
+
+        let ex = expectation(description: "download-populated-realm-async-open")
+        autoOpen(appId: flexibleSyncAppId, openType: .configuration(configuration)) { autoOpenState in
+            if case let .open(realm) = autoOpenState {
+                XCTAssertNotNil(realm)
+                switch configuration.syncConfiguration!.clientResetMode {
+                case .manual(let block):
+                    XCTAssertNotNil(block)
+                default:
+                    XCTFail("Should be set to manual")
+                }
+                ex.fulfill()
             }
         }
     }