+++ RealmSwift/SwiftUI.swift
@@ -1526,24 +1526,36 @@ private class ObservableAsyncOpenStorage: ObservableObject {
     }
 
     private func asyncOpenForUser(_ user: User) {
-        // Set the `syncConfiguration` depending if there is partition value (pbs) or not (flx).
         var config: Realm.Configuration
-        if let partitionValue = partitionValue {
-            config = user.configuration(partitionValue: partitionValue, cancelAsyncOpenOnNonFatalErrors: true)
+
+        if let configuration = self.configuration {
+            config = configuration
+            var syncConfig: SyncConfiguration?
+            if let syncConfiguration = configuration.syncConfiguration {
+                if let partitionValue = partitionValue {
+                    syncConfig = user.configuration(partitionValue: partitionValue, clientResetMode: syncConfiguration.clientResetMode).syncConfiguration
+                }
+            } else {
+                if let partitionValue = partitionValue {
+                    syncConfig = user.configuration(partitionValue: partitionValue).syncConfiguration
+                } else {
+                    syncConfig = user.flexibleSyncConfiguration().syncConfiguration
+                }
+            }
+            if let syncConfig = syncConfig {
+                config.syncConfiguration = syncConfig
+            }
         } else {
-            config = user.flexibleSyncConfiguration(cancelAsyncOpenOnNonFatalErrors: true)
+            if let partitionValue = partitionValue {
+                config = user.configuration(partitionValue: partitionValue)
+            } else {
+                config = user.flexibleSyncConfiguration()
+            }
         }
 
-        // Use the user configuration by default or set configuration with the current user `syncConfiguration`'s.
-        if var configuration = configuration {
-            // We want to throw if the configuration doesn't contain a `SyncConfiguration`
-            guard configuration.syncConfiguration != nil else {
-                throwRealmException("The used configuration was not configured with sync.")
-            }
-            let userSyncConfig = config.syncConfiguration
-            configuration.syncConfiguration = userSyncConfig
-            config = configuration
-        }
+        let syncConfiguration = config.syncConfiguration
+        syncConfiguration?.config.cancelAsyncOpenOnNonFatalErrors = true
+        config.syncConfiguration = syncConfiguration
 
         // Cancel any current subscriptions to asyncOpen if there is one
         cancelAsyncOpen()
@@ -1585,7 +1597,7 @@ private class ObservableAsyncOpenStorage: ObservableObject {
         }
 
         // We don't want to use the `defaultConfiguration` from the environment, we only want to use this environment value in @AsyncOpen if is not the default one
-        if configuration != .defaultConfiguration, self.configuration != configuration {
+        if configuration == .defaultConfiguration, self.configuration != configuration {
             if let partitionValue = configuration.syncConfiguration?.partitionValue {
                 self.partitionValue = partitionValue
             }
@@ -1697,14 +1709,15 @@ private class ObservableAsyncOpenStorage: ObservableObject {
     }
 
     /**
-     Initialize the property wrapper
+     Initialize the property wrapper for a given configuration or Partition.
      - parameter appId: The unique identifier of your Realm app, if empty or `nil` will try to retrieve latest singular cached app.
      - parameter partitionValue: The `BSON` value the Realm is partitioned on.
-     - parameter configuration: The `Realm.Configuration` used when creating the Realm,
-                 user's sync configuration for the given partition value will be set as the `syncConfiguration`,
-                 if empty the user configuration will be used.
+     - parameter configuration: A configuration `Realm.configuration` to use when opening the Realm.
      - parameter timeout: The maximum number of milliseconds to allow for a connection to
                  become fully established., if empty or `nil` no connection timeout is set.
+
+     - note: This intializer will use either the configuration (if not nil) with the given partition value, or build a configuration from the
+             partition value using `user.configuration(partitionValue:)`.
      */
     public init<Partition>(appId: String? = nil,
                            partitionValue: Partition,
@@ -1716,13 +1729,14 @@ private class ObservableAsyncOpenStorage: ObservableObject {
     }
 
     /**
-     Initialize the property wrapper for a flexible sync configuration.
+     Initialize the property wrapper for a flexible sync app.
      - parameter appId: The unique identifier of your Realm app, if empty or `nil` will try to retrieve latest singular cached app.
-     - parameter configuration: The `Realm.Configuration` used when creating the Realm,
-                 user's sync configuration for the given partition value will be set as the `syncConfiguration`,
-                 if empty the user configuration will be used.
+     - parameter configuration: A configuration `Realm.configuration` to use when opening the Realm.
      - parameter timeout: The maximum number of milliseconds to allow for a connection to
                  become fully established., if empty or `nil` no connection timeout is set.
+
+     - note: This intializer will use either the configuration (if not nil), or build a configuration from the
+             partition value using `user.flexibleSyncConfiguration()`.
      */
     public init(appId: String? = nil,
                 configuration: Realm.Configuration? = nil,
@@ -1809,14 +1823,14 @@ private class ObservableAsyncOpenStorage: ObservableObject {
     }
 
     /**
-     Initialize the property wrapper
-     - parameter appId: The unique identifier of your Realm app,  if empty or `nil` will try to retrieve latest singular cached app.
+     Initialize the property wrapper for a given configuration or Partition.
+     - parameter appId: The unique identifier of your Realm app, if empty or `nil` will try to retrieve latest singular cached app.
      - parameter partitionValue: The `BSON` value the Realm is partitioned on.
-     - parameter configuration: The `Realm.Configuration` used when creating the Realm,
-                 user's sync configuration for the given partition value will be set as the `syncConfiguration`,
-                 if empty the user configuration will be used.
+     - parameter configuration: A configuration `Realm.configuration` to use when opening the Realm.
      - parameter timeout: The maximum number of milliseconds to allow for a connection to
-                 become fully established, if empty or `nil` no connection timeout is set.
+                 become fully established., if empty or `nil` no connection timeout is set.
+
+     - note: This intialiser will use either the configuration or build a configuration from the partition value using `user.configuration(partitionValue:)`.
      */
     public init<Partition>(appId: String? = nil,
                            partitionValue: Partition,
@@ -1828,11 +1842,9 @@ private class ObservableAsyncOpenStorage: ObservableObject {
     }
 
     /**
-     Initialize the property wrapper for a flexible sync configuration.
+     Initialize the property wrapper with a configuration.
      - parameter appId: The unique identifier of your Realm app, if empty or `nil` will try to retrieve latest singular cached app.
-     - parameter configuration: The `Realm.Configuration` used when creating the Realm,
-                 user's sync configuration for the given partition value will be set as the `syncConfiguration`,
-                 if empty the user configuration will be used.
+     - parameter configuration: A configuration `Realm.configuration` to use when opening the Realm.
      - parameter timeout: The maximum number of milliseconds to allow for a connection to
                  become fully established., if empty or `nil` no connection timeout is set.
      */